+++ src/pycture/commands/edit_commands.py
@@ -1,0 +1,10 @@
+from PyQt5.QtWidgets import QWidget, QMainWindow
+
+from .command import Command
+
+class EditBrightnessCommand(Command):
+    def __init__(self, parent: QWidget):
+        super().__init__("Bright", parent)
+
+    def execute(self, main_window: QMainWindow):
+        print("Edits brightness")
